Auto-Update gpu_gv-algo.sh from GPU-skeleton directory and optimized code into a GRID loop
1. If gpu_gv-algo.sh is advanced or changed in the GPU-skeleton directory,
then all the other gpu_gv-algo.sh scripts in the GPU-UUID directories
update themselves before the next calculation run (every 31 seconds)
or at startup

2. Instead of having three variables and three for-loops to process them, we use an indexed array GRID[] with the 3 members "netz", "solar" and "solar_akku".
The values of this GRID[] are then used as keys/index of the associative arrays kwh_BTC[] and kwh_EUR[].
The names of the "kwh_GRID[]_kosten.in" files as well as the names of the "best_algo_GRID.out" files are also derived from the GRID[] array.
This ends up in a for-loop through GRID[], calculating the best_algos instead of 3 nearly identical code blocks, which calculated the best_algos.